build: configure project to build with cmake
This revision features some pretty significant improvements to the build
infrastructure for opentracker. The project has been reworked and
modernized with CMake to enhance the customizability and flexiblity of
builds. Overall, opentracker should now be easier to build across
different environments.

The new build script is functionally equivalent with the makefiles it
replaces. All existing compiler definitions for build features are
maintained, but can now be configured from the outside during the
configuration phase.

A significant amount of effort went into the integration with libowfat.
The new build supports both static/dynamic linking to the library
(selectable with WANT_STATIC_LIBOWFAT). The find module for libowfat
also attempts to read the library version so that WANT_NO_AUTO_FREE can
be enabled automatically if needed, although this is done in a best
effort manner because the library doesn't expose a clean way to
determine the version (upstream Debian bug has been filed). Opentracker
can be built against a system-wide installation of libowfat (e.g. from
debian repositories) or a source build.

An INSTALL doc has been added to document the new build. It offers a
quick summary on how to build opentracker in debug/release
configurations, notes on dependencies and the different build options
supported.

It appears the proxy has been broken for some time, so I've excluded it
from the built. We can uncomment the build script when fixed.

I've kept the original build scripts for now. We can decide to remove
them later if appropriate.

Changes have been tested on Debian bookworm/bullseye/trixie, Ubuntu
24.04/22.04, Fedora 42. Can't seem to build libowfat on Arch. Haven't
tested on BSD.